acmetool: select the last most thumbprint
This makes things work where there are two thumbprints on a system -
e.g. acmev1 and acmev2. Select the last most (i.e. the most recent
thumbprint) instead of the first (i.e. oldest) thumbprint.

Note: the list is split on a word boundary, and is formatted

<thumbprint> <url>

so to select the last thumbprint, we select the second-to-last word in
the string.